Rename package utils to core and REGENERATE package-lock.json!!
