+++ backend/src/core/dynamodb_characters.ts
