+++ backend/src/core/dynamodb_client.ts
