+++ backend/src/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "utils",
+  "name": "core",
   "version": "1.0.0",
   "main": "index.mjs",
   "type": "module",
