+++ backend/src/lambdas/add-history-record/index.ts
@@ -24,7 +24,7 @@ import {
   logAndEnsureHttpError,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 import {
   AddHistoryRecordRequest,
   AddHistoryRecordResponse,
