+++ backend/src/lambdas/add-special-ability/index.ts
@@ -17,7 +17,7 @@ import {
   setSpecialAbilities,
   logZodError,
   isZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return _addSpecialAbility({
