+++ backend/src/lambdas/clone-character/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "dependencies": {
     "api-spec": "*",
-    "utils": "*"
+    "core": "*"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.145",
