+++ backend/src/lambdas/create-character/index.ts
@@ -10,7 +10,7 @@
 //   createCharacterItem,
 //   isZodError,
 //   logZodError,
-// } from "utils";
+// } from "core";
 // import {
 //   dis_advantagesSchema,
 //   generalInformationSchema,
