+++ backend/src/lambdas/create-character/package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "config": "*",
     "api-spec": "*",
-    "utils": "*"
+    "core": "*"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.145",
