+++ backend/src/lambdas/get-history/index.ts
@@ -19,7 +19,7 @@ import {
   decodeUserId,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return getHistory({
