+++ backend/src/lambdas/get-skill-increase-cost/index.ts
@@ -18,7 +18,7 @@ import {
   logAndEnsureHttpError,
   logZodError,
   isZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return getSkillCost({
