+++ backend/src/lambdas/revert-history-record/package.json
@@ -4,7 +4,7 @@
   "main": "index.mjs",
   "type": "module",
   "dependencies": {
-    "utils": "*",
+    "core": "*",
     "api-spec": "*"
   },
   "devDependencies": {
