+++ backend/src/lambdas/set-history-comment/index.ts
@@ -22,7 +22,7 @@ import {
   logZodError,
   isZodError,
   getCharacterItem,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return setHistoryComment({
