+++ backend/src/lambdas/update-attribute/index.ts
@@ -24,7 +24,7 @@ import {
   updateBaseValue,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return _updateAttribute({
