+++ backend/src/lambdas/update-base-value/index.ts
@@ -22,7 +22,7 @@ import {
   updateBaseValue,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return _updateBaseValue({
