+++ backend/src/lambdas/update-calculation-points/index.ts
@@ -21,7 +21,7 @@ import {
   updateAttributePoints,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return _updateCalculationPoints({
