+++ backend/src/lambdas/update-combat-values/index.ts
@@ -20,7 +20,7 @@ import {
   updateCombatValues,
   isZodError,
   logZodError,
-} from "utils";
+} from "core";
 
 export const handler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   return _updateCombatValues({
