+++ backend/src/lambdas/update-level/package.json
@@ -4,7 +4,7 @@
   "main": "index.mjs",
   "type": "module",
   "dependencies": {
-    "utils": "*",
+    "core": "*",
     "api-spec": "*"
   },
   "devDependencies": {
