+++ package.json
@@ -8,7 +8,7 @@
     "scripts",
     "api-spec",
     "backend/src/config",
-    "backend/src/utils",
+    "backend/src/core",
     "backend/src/lambdas/*"
   ],
   "scripts": {
